# openapi.json
{"openapi":"3.1.0","info":{"title":"HeyPico Maps API","description":"Google Maps integration for Open WebUI with local LLM","version":"1.0.0"},"paths":{"/api/maps/search":{"post":{"tags":["maps"],"summary":"Search for places","description":"Search for places using Google Maps Places API","operationId":"search_places_api_maps_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceSearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful search","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceSearchResponse"}}}},"400":{"description":"Invalid request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/maps/place/{place_id}":{"get":{"tags":["maps"],"summary":"Get place details","description":"Get detailed information about a specific place","operationId":"get_place_details_api_maps_place__place_id__get","parameters":[{"name":"place_id","in":"path","required":true,"schema":{"type":"string","title":"Place Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceDetailsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/maps/directions":{"post":{"tags":["maps"],"summary":"Get directions","description":"Get turn-by-turn directions between two locations","operationId":"get_directions_api_maps_directions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectionsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectionsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/maps/geocode":{"post":{"tags":["maps"],"summary":"Geocode address","description":"Convert an address to geographic coordinates","operationId":"geocode_address_api_maps_geocode_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeocodeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeocodeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health Check","description":"Health check endpoint.\n\nReturns service status and configuration information.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}}},"components":{"schemas":{"DirectionsRequest":{"properties":{"origin":{"type":"string","maxLength":200,"minLength":1,"title":"Origin","description":"Starting point (address or place name)"},"destination":{"type":"string","maxLength":200,"minLength":1,"title":"Destination","description":"Ending point (address or place name)"},"mode":{"type":"string","title":"Mode","description":"Travel mode: driving, walking, bicycling, transit","default":"driving"}},"type":"object","required":["origin","destination"],"title":"DirectionsRequest","description":"Request model for directions."},"DirectionsResponse":{"properties":{"origin":{"type":"string","title":"Origin"},"destination":{"type":"string","title":"Destination"},"mode":{"type":"string","title":"Mode"},"route":{"$ref":"#/components/schemas/DirectionsRoute"},"google_maps_url":{"type":"string","title":"Google Maps Url"}},"type":"object","required":["origin","destination","mode","route","google_maps_url"],"title":"DirectionsResponse","description":"Response model for directions."},"DirectionsRoute":{"properties":{"summary":{"type":"string","title":"Summary","description":"Route summary"},"distance":{"type":"string","title":"Distance","description":"Total distance"},"duration":{"type":"string","title":"Duration","description":"Total duration"},"start_address":{"type":"string","title":"Start Address","description":"Starting address"},"end_address":{"type":"string","title":"End Address","description":"Ending address"},"steps":{"items":{"$ref":"#/components/schemas/DirectionsStep"},"type":"array","title":"Steps","description":"Turn-by-turn steps"}},"type":"object","required":["summary","distance","duration","start_address","end_address","steps"],"title":"DirectionsRoute","description":"Directions route information."},"DirectionsStep":{"properties":{"instruction":{"type":"string","title":"Instruction","description":"HTML instructions"},"distance":{"type":"string","title":"Distance","description":"Distance text (e.g., '0.5 km')"},"duration":{"type":"string","title":"Duration","description":"Duration text (e.g., '5 mins')"}},"type":"object","required":["instruction","distance","duration"],"title":"DirectionsStep","description":"Individual direction step."},"ErrorResponse":{"properties":{"error":{"type":"string","title":"Error","description":"Error message"},"detail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Detail","description":"Detailed error information"},"status_code":{"type":"integer","title":"Status Code","description":"HTTP status code"}},"type":"object","required":["error","status_code"],"title":"ErrorResponse","description":"Standard error response."},"GeocodeLocation":{"properties":{"formatted_address":{"type":"string","title":"Formatted Address"},"location":{"$ref":"#/components/schemas/PlaceLocation"},"location_type":{"type":"string","title":"Location Type"},"place_id":{"type":"string","title":"Place Id"}},"type":"object","required":["formatted_address","location","location_type","place_id"],"title":"GeocodeLocation","description":"Geocoding result location."},"GeocodeRequest":{"properties":{"address":{"type":"string","maxLength":300,"minLength":1,"title":"Address","description":"Address to geocode"}},"type":"object","required":["address"],"title":"GeocodeRequest","description":"Request model for geocoding."},"GeocodeResponse":{"properties":{"address":{"type":"string","title":"Address","description":"Original address query"},"results":{"items":{"$ref":"#/components/schemas/GeocodeLocation"},"type":"array","title":"Results","description":"Geocoding results"},"count":{"type":"integer","title":"Count","description":"Number of results"}},"type":"object","required":["address","results","count"],"title":"GeocodeResponse","description":"Response model for geocoding."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","description":"Service status"},"service":{"type":"string","title":"Service","description":"Service name"},"version":{"type":"string","title":"Version","description":"Service version"},"maps_api_configured":{"type":"boolean","title":"Maps Api Configured","description":"Whether Google Maps API key is configured"}},"type":"object","required":["status","service","version","maps_api_configured"],"title":"HealthResponse","description":"Health check response."},"PlaceDetailsResponse":{"properties":{"name":{"type":"string","title":"Name"},"formatted_address":{"type":"string","title":"Formatted Address"},"formatted_phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Formatted Phone Number"},"international_phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"International Phone Number"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"user_ratings_total":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"User Ratings Total"},"price_level":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Price Level"},"opening_hours":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Opening Hours"},"location":{"$ref":"#/components/schemas/PlaceLocation"},"types":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Types"},"google_maps_url":{"type":"string","title":"Google Maps Url"}},"type":"object","required":["name","formatted_address","location","google_maps_url"],"title":"PlaceDetailsResponse","description":"Response model for place details."},"PlaceLocation":{"properties":{"lat":{"type":"number","title":"Lat","description":"Latitude"},"lng":{"type":"number","title":"Lng","description":"Longitude"}},"type":"object","required":["lat","lng"],"title":"PlaceLocation","description":"Location coordinates."},"PlaceResult":{"properties":{"name":{"type":"string","title":"Name","description":"Place name"},"address":{"type":"string","title":"Address","description":"Formatted address"},"place_id":{"type":"string","title":"Place Id","description":"Google Maps Place ID"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating","description":"Average rating (0-5)"},"user_ratings_total":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"User Ratings Total","description":"Number of ratings"},"location":{"$ref":"#/components/schemas/PlaceLocation","description":"Geographic coordinates"},"types":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Types","description":"Place types"},"google_maps_url":{"type":"string","title":"Google Maps Url","description":"Google Maps URL"}},"type":"object","required":["name","address","place_id","location","google_maps_url"],"title":"PlaceResult","description":"Individual place search result."},"PlaceSearchRequest":{"properties":{"query":{"type":"string","maxLength":200,"minLength":1,"title":"Query","description":"Search query (e.g., 'pizza restaurants', 'gas stations')"},"location":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Location","description":"Center location for search (e.g., 'New York, NY')"},"radius":{"type":"integer","maximum":50000.0,"minimum":1.0,"title":"Radius","description":"Search radius in meters (1-50000)","default":5000}},"type":"object","required":["query"],"title":"PlaceSearchRequest","description":"Request model for place search."},"PlaceSearchResponse":{"properties":{"query":{"type":"string","title":"Query","description":"Original search query"},"results":{"items":{"$ref":"#/components/schemas/PlaceResult"},"type":"array","title":"Results","description":"List of matching places"},"count":{"type":"integer","title":"Count","description":"Number of results returned"}},"type":"object","required":["query","results","count"],"title":"PlaceSearchResponse","description":"Response model for place search."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}